tests/all: rename test files, change load order of test files in mocharc
